Add custom hooks for tasks and favorites functions
