+++ src/hooks/useFavorites.js
@@ -1,0 +1,34 @@
+import { useEffect, useState } from 'react'
+import { loadFromLocal, saveToLocal } from '../utils/localStorage'
+
+export default function useFavorites() {
+  const [favoriteCrops, setFavoriteCrops] = useState(
+    loadFromLocal('favoriteCrops') ?? []
+  )
+
+  useEffect(() => {
+    saveToLocal('favoriteCrops', favoriteCrops)
+  }, [favoriteCrops])
+
+  function toggleFavorite(crop) {
+    const isInFavorites = favoriteCrops.some(favCrop => favCrop.id === crop.id)
+    if (isInFavorites) {
+      removeFromFavorites(crop.id)
+    } else {
+      addToFavorites(crop)
+    }
+  }
+
+  function addToFavorites(crop) {
+    setFavoriteCrops([...favoriteCrops, crop])
+  }
+
+  function removeFromFavorites(id) {
+    setFavoriteCrops(favoriteCrops.filter(favCrop => favCrop.id !== id))
+  }
+
+  return {
+    toggleFavorite,
+    favoriteCrops,
+  }
+}
